Updated data types

+++ Binary_Search/Silver_P1_2020-Social_Distancing/SocDist.java
@@ -43,16 +43,16 @@ import java.util.*;
 // Binary search on multiple answers method
 // T: O(N log D + M log M) = O((N+M)log(maxDist)), M: O(M), where N is the number of cows, M is the number of intervals, and D is the range of distance
 public class SocDist {
-    private static long[][] intervals;  // make intervals accessible by all methods
+    private static int[][] intervals;  // make intervals accessible by all methods
     private static int n, m;  // number of cows and intervals respectively
 
     // method to test if a certain minimum distance can be achieved between cows
-    private static boolean possible_placement(long min_dist){
+    private static boolean possible_placement(int min_dist){
         // initialize previous cow location to be right before the first position
-        long prev_cow_loc = intervals[0][0] - min_dist;
+        int prev_cow_loc = intervals[0][0] - min_dist;
         int remaining_cows = n;
-        for (long[] interval : intervals){
-            long start = interval[0], end = interval[1];
+        for (int[] interval : intervals){
+            int start = interval[0], end = interval[1];
             // if cow placement is not in the grassy area, update prev_cow_loc right before next grassy area
             if (prev_cow_loc + min_dist < start)
                 prev_cow_loc = start - min_dist;
@@ -74,23 +74,23 @@ public class SocDist {
         n = Integer.parseInt(st.nextToken());
         m = Integer.parseInt(st.nextToken());
         // Create a 2D array of all the intervals in the next m lines
-        intervals = new long[m][2];
+        intervals = new int[m][2];
         for (int interval = 0; interval < m; interval++){
             // use new StringTokenizer for each line
             st = new StringTokenizer(in.readLine());
-            intervals[interval][0] = Long.parseLong(st.nextToken());  // start point
-            intervals[interval][1] = Long.parseLong(st.nextToken());  // end point
+            intervals[interval][0] = Integer.parseInt(st.nextToken());  // start point
+            intervals[interval][1] = Integer.parseInt(st.nextToken());  // end point
         }
         in.close();
 
         // Only sort intervals by start point because there are no overlapping intervals
-        Arrays.sort(intervals, Comparator.comparingLong(a -> a[0]));
+        Arrays.sort(intervals, Comparator.comparingInt(a -> a[0]));
 
         // Use Binary search algorithm to quickly test multiple solutions
         // worst case = 1, best case = equally distant cows
-        long lo = 1, hi = (intervals[m-1][1] - intervals[0][0]) / (n - 1);
+        int lo = 1, hi = (intervals[m-1][1] - intervals[0][0]) / (n - 1);
         while (lo < hi){
-            long mid = lo + (hi - lo + 1) / 2;
+            int mid = lo + (hi - lo + 1) / 2;
             // create an algorithm that tests out this minimum distance
             if (possible_placement(mid))
                 lo = mid;
